Add predicateBuilder to deal with predicate related from MySQL dialect to generic parse module
